update: changed IntNode to accept values in range int64_t

+++ linked_list.h
@@ -5,31 +5,31 @@ typedef struct IntNode IntNode;
 struct IntNode{
 	IntNode* previous;
 	IntNode* next;
-	uint32_t value;
+	int64_t value;
 };
 
 typedef struct {
 	IntNode* start;
 	IntNode* center;
 	IntNode* end;
-	uint32_t size;
+	int64_t size;
 } IntLinkedList;
 
-IntLinkedList* InitIntLinkedList(uint32_t values[], int length);
+IntLinkedList* InitIntLinkedList(int64_t values[], int length);
 
 void DestroyIntLinkedList(IntLinkedList* linkedList);
 
-uint32_t RemoveIntNodeAt(IntLinkedList* linked, uint32_t idx);
-void PutIntNodeAt(IntLinkedList* linked, uint32_t value, uint32_t idx);
+int64_t RemoveIntNodeAt(IntLinkedList* linked, int64_t idx);
+void PutIntNodeAt(IntLinkedList* linked, int64_t value, int64_t idx);
 
 // removes at first position
-uint32_t PopIntNode(IntLinkedList* linkedList);
+int64_t PopIntNode(IntLinkedList* linkedList);
 // inserts at first position
-void PushIntNode(IntLinkedList* linkedList, uint32_t value);
+void PushIntNode(IntLinkedList* linkedList, int64_t value);
 
-void AppendIntNode(IntLinkedList* linkedList, uint32_t value);
+void AppendIntNode(IntLinkedList* linkedList, int64_t value);
 
-uint32_t RemoveTail(IntLinkedList* linkedList);
+int64_t RemoveTail(IntLinkedList* linkedList);
 
 void PrintIntLinkedList(IntLinkedList* linked);
 void PrintIntLinkedListReversed(IntLinkedList* linked);
